Made Article component that is reusable few touches regarding response
