+++ src/UI/AppLayout.jsx
@@ -1,12 +1,7 @@
-import styled from 'styled-components';
 import Header from '../components/Header';
 import { Outlet } from 'react-router-dom';
 import { useState } from 'react';
 import NewsSourceModal from './NewsSourceModal';
-
-const StyledAppLayout = styled.div`
-  min-height: 100vh;
-`;
 
 function AppLayout() {
   const [isModalOpen, setIsModalOpen] = useState(false);
@@ -16,11 +11,13 @@ function AppLayout() {
   };
 
   return (
-    <StyledAppLayout>
-      <Header toggleModal={toggleModal} />
+    <div className="box-border w-full">
+      <Header toggleModal={toggleModal} className="sticky top-0" />
+
       <NewsSourceModal isOpen={isModalOpen} toggleModal={toggleModal} />
-      <Outlet />
-    </StyledAppLayout>
+
+      <Outlet className="min-h-screen min-w-full" />
+    </div>
   );
 }
