+++ src/components/Header.jsx
@@ -1,25 +1,10 @@
-import styled from 'styled-components';
-import SearchBar from '../UI/SearchBar';
+import SearchAndPreferences from '../UI/SearchAndPreferences';
 import NewspaperIcon from '@mui/icons-material/Newspaper';
 import { Link } from 'react-router-dom';
 
-const StyledHeader = styled.header`
-  background-color: rgba(255, 255, 255, 0.5);
-  backdrop-filter: blur(10px);
-  -webkit-backdrop-filter: blur(10px);
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-  padding: 1.2rem 3rem;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  top: 0;
-  position: sticky;
-  z-index: 100;
-`;
-
 function Header({ toggleModal }) {
   return (
-    <StyledHeader>
+    <div className="sticky top-0 z-50 flex items-center justify-between border-b border-black/10 bg-white/50 px-2 py-5 backdrop-blur-md">
       <Link
         to="/"
         style={{
@@ -30,10 +15,10 @@ function Header({ toggleModal }) {
         }}
       >
         <NewspaperIcon fontSize="large" />
-        <div>Global Pulse</div>
+        <div className="hidden md:block">Global Pulse</div>
       </Link>
-      <SearchBar toggleModal={toggleModal} />
-    </StyledHeader>
+      <SearchAndPreferences toggleModal={toggleModal} />
+    </div>
   );
 }
