+++ src/components/SearchArticles.jsx
@@ -1,19 +1,10 @@
-import styled from "styled-components";
-import { Search } from "../UI/Search";
-
-const ActionBarStyled = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  padding: 1rem;
-  align-items: center;
-`;
+import { SearchNews } from '../UI/SearchNews';
 
 function FilterAndFind() {
   return (
-    <ActionBarStyled>
-      <Search />
-    </ActionBarStyled>
+    <div className="w-full p-1">
+      <SearchNews />
+    </div>
   );
 }
