+++ src/main.jsx
@@ -1,12 +1,13 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App.jsx";
-import { SearchProvider } from "./service/SearchContext.jsx";
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import App from './App.jsx';
+import './index.css';
+import { SearchProvider } from './service/SearchContext.jsx';
 
-ReactDOM.createRoot(document.getElementById("root")).render(
+ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
     <SearchProvider>
       <App />
     </SearchProvider>
-  </React.StrictMode>
+  </React.StrictMode>,
 );
