+++ src/picture/deluge-2.1.1-win64-setup.exe.sha256
@@ -1,0 +1,1 @@
+183f401f088cbb973633dbeeb16ded1326b7aa567acfc9de32855d5be14653ae  deluge-2.1.1-win64-setup.exe
